requirement.txt
pythonnet==3.0.5
PySide6==6.10.0
pyyaml==6.0.3 
psutil=7.1.3
pyinstaller>=6.0.0
